requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
Django==3.0.3
future==0.18.2
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
python-telegram-bot==12.4.2
pytz==2019.3
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
tornado==6.0.4
